requirements.txt
asgiref==3.8.1
Django==5.0.3
django-admin-autocomplete-filter==0.7.1
django-allauth==0.62.1
django-allauth-ui==0.2.2
django-dynamic-admin-forms==3.0.0
django-model-utils==4.5.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-tinymce==4.0.0
django-unfold==0.22.0
django-widget-tweaks==1.5.0
gunicorn==21.2.0
jsmin==3.0.1
packaging==24.0
pillow==10.3.0
pycountry==23.12.11
sqlparse==0.4.4
whitenoise==6.6.0
